+++ packages/contracts/contracts/registry/RegistryUtils.sol
@@ -1,0 +1,34 @@
+// SPDX-License-Identifier: MIT
+
+pragma solidity 0.8.10;
+
+/// @notice Checks if the name is either 0-9, a-z or a dash (-).
+/// @param subDomain The name of the DAO.
+/// @return `true` if the name is valid or `false` if at least one char is invalid.
+/// @dev Aborts on the first invalid char found.
+function isSubdomainValid(string calldata subDomain) pure returns (bool) {
+    bytes calldata nameBytes = bytes(subDomain);
+    uint256 nameLength = nameBytes.length;
+    for (uint256 i; i < nameLength; i++) {
+        uint8 char = uint8(nameBytes[i]);
+
+        // if char is between a-z
+        if (char > 96 && char < 123) {
+            continue;
+        }
+
+        // if char is between 0-9
+        if (char > 47 && char < 58) {
+            continue;
+        }
+
+        // if char is -
+        if (char == 45) {
+            continue;
+        }
+
+        // invalid if one char doesn't work with the rules above
+        return false;
+    }
+    return true;
+}