+++ packages/contracts/deploy/30_plugins/10_plugin-repos/00_create-plugins-repo.ts
@@ -32,7 +32,7 @@ const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
   );
   await createPluginRepo(
     hre,
-    'AddresslistVoting',
+    'address-list-voting',
     'AddresslistVotingSetup',
     ethers.utils.hexlify(
       ethers.utils.toUtf8Bytes(`ipfs://${addresslistReleaseCIDPath}`)
@@ -53,7 +53,7 @@ const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
   );
   await createPluginRepo(
     hre,
-    'TokenVoting',
+    'token-voting',
     'TokenVotingSetup',
     ethers.utils.hexlify(
       ethers.utils.toUtf8Bytes(`ipfs://${tokenReleaseCIDPath}`)
@@ -74,7 +74,7 @@ const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
   );
   await createPluginRepo(
     hre,
-    'Admin',
+    'admin',
     'AdminSetup',
     ethers.utils.hexlify(
       ethers.utils.toUtf8Bytes(`ipfs://${adminReleaseCIDPath}`)
@@ -95,7 +95,7 @@ const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
   );
   await createPluginRepo(
     hre,
-    'Multisig',
+    'multisig',
     'MultisigSetup',
     ethers.utils.hexlify(
       ethers.utils.toUtf8Bytes(`ipfs://${multisigReleaseCIDPath}`)