+++ packages/subgraph/src/packages/admin/admin.ts
@@ -90,6 +90,7 @@ export function handleProposalExecuted(event: ProposalExecuted): void {
   let proposalEntity = AdminProposal.load(proposalId);
   if (proposalEntity) {
     proposalEntity.executed = true;
+    proposalEntity.executionTxHash = event.transaction.hash;
     proposalEntity.save();
 
     // update actions