+++ packages/subgraph/CHANGELOG.md
@@ -9,11 +9,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added `executionTxHash` to `AddresslistVotingProposal`, `TokenVotingProposal`, `MultisigProposal` and `AdminProposal`.
 - Added `voteReplaced` and `updatedAt` to `TokenVotingVote` and `AddresslistVotingVote`.
 
 ### Changed
 
-- Changes callId in `Executed` event from `uint256` to `bytes32`
+- Changes `callId` in `Executed` event from `uint256` to `bytes32`.
+- Renames `name` in `PluginRepoRegistry` and `DAORegistry` to `subdomain`.
+
 
 ## 0.6.2-alpha